Fix the client-side problem described in issue #2382 (correctly handle dual-stack servers in ra_svn client).
With this fix, clients iterate over all available addresses when
connecting to a host running svnserve and use the first address
that works.

We used to only try the first address, and if that didn't work
we bailed out -- too bad if want to connect to localhost, the
first address happens to be IPv6 loopback, and the host is
configured to reject any IPv6 packets.
See http://svn.haxx.se/dev/archive-2008-03/0605.shtml

Patch by: joe
(Updated for current trunk by me.)

Approved by: lgo

* subversion/libsvn_ra_svn/client.c:
  (make_connection): Iterate through the returned list of addresses
   until a socket/connect pair succeeds.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@870078 13f79535-47bb-0310-9956-ffa450edef68