Follow-up to r30145, remove some now unused struct fields in merge.c.
Previously we needed access to a pool with a lifetime as long as
do_merge.c()'s iteration over multiple merge sources.  But this was *only* 
needed to support not recording mergeinfo for no-op merges, which was 
removed in r30145, so we can safely remove the "long lived pool" we 
passed down inside the merge command baton. 

* subversion/libsvn_client/merge.c
  (merge_cmd_baton_t, get_mergeinfo_walk_baton): Remove long_pool fields from
  both of these structs.
  (get_mergeinfo_paths): Remove initialization of long_pool fields in
  get_mergeinfo_walk_baton wb.
  (do_merge): Remove initialization of long_pool field in merge_cmd_baton_t.

Suggested by: julianfoad

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@870259 13f79535-47bb-0310-9956-ffa450edef68