Fix Issue #3158. Merged deletions causing non-inheritable mergeinfo in subsequent changes.
When merging to a target which has missing subtrees, our merge tracking code
puts explicit mergeinfo on the missing subtrees parent and siblings.  The
parent gets non-inheritable ranges describing the merge and the siblings
get normal ranges.  This is all to prevent the missing subtree from 
inheriting incorrect mergeinfo.  This change *stops* categorizing subtrees 
scheduled for deletion as missing and so suppresses the aforementioned
behavior for deleted paths.

* subversion/libsvn_client/merge.c
  (insert_parent_and_sibs_of_sw_absent_del_entry): Don't process paths
  scheduled for deletion anymore.
  (get_mergeinfo_paths): Tweak doc string and comments to reflect the above.

* subversion/tests/cmdline/merge_tests.py
  (mergeinfo_recording_in_skipped_merge): Tweak test expectations to line up
  with new behavior.
  (test_list): Remove XFail from merge_loses_mergeinfo test.  The only reason
  this test has been failing is because the old behavior regarding scheduled
  for deletion paths was not expected.  Now that that behavior is gone, the
  test passes.

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@870323 13f79535-47bb-0310-9956-ffa450edef68