Make --record-only merges record and elide mergeinfo on subtrees.
Previously, --record-only merge only set mergeinfo on the merge
target, ignored subtrees with explicit mergeinfo, and did not attempt
any elision.  Basically, a "smart propedit" on the the merge target.

After this change, it's still a smart propedit, but it also notices
subtrees with explicit mergeinfo and sets mergeinfo on them as if we
had done a "real" merge, and then it attempts to elide those subtrees'
mergeinfo if possible (again like a "real" merge).

* subversion/libsvn_client/merge.c
  (populate_remaining_ranges): Expect that this code may now be invoked
  during --record-only merges and in that case follow the same code path
  we do when not honoring mergeinfo, i.e. the requested range *is* the 
  remaining range, no filtering needed.
  (record_mergeinfo_for_record_only_merge): Delete, a few minor tweaks to
  do_file_merge(), do_directory_merge(), and their helpers is adequate to
  replace this while also implementing the new --record-only functionality.  
  (do_file_merge, do_directory_merge): Expect that this code may now be 
  invoked during --record-only merges; if it is, *skip* the actual
  merging, but *do* record the merge in mergeinfo.
  (do_merge): Remove call to record_mergeinfo_for_record_only_merge().


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@870718 13f79535-47bb-0310-9956-ffa450edef68