Fix yet another issue #3067 bug.
When driving the merge report editor and the merge target has subtrees that
need different ranges applied, there was a case we didn't handle: A subtree
may need a range merged that doesn't intersect with the range its nearest
parent requires *nor* that any other subtree requires.  In some cases this
could cause us to describe path/revisions to the reporter that don't exist
resulting in the dreaded "svn: Working copy path 'blah' does not exist in
repository" errors.

* subversion/libsvn_client/merge.c
  (drive_merge_report_editor): When merging a range X:Y if we find a subtree
  that doesn't need any part of that range merged then don't describe that
  subtree to the reporter.

* subversion/tests/cmdline/merge_tests.py
  (test_list): Remove XFail from
  merge_target_and_subtrees_need_nonintersecting_ranges.



git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@872781 13f79535-47bb-0310-9956-ffa450edef68