Two changes committed at once. Reverted later in r33834, to be re-applied separately.
Introduce a macro "U_()" for marking messages that are not useful for
translation, to indicate that a conscious decision was made not to translate
them.

* subversion/svn_private_config.hw,
  configure.ac:
  (U_): New macro.

* www/hacking.html
  Document U_().

* notes/l10n-problems
  Tweak to acknowledge that not all error messages are to be translated.

Change some internal error detection code to use the generic SVN_ERR_ASSERT
and SVN_ERR_MALFUNCTION macros instead of specific messages, because the
user won't benefit from knowing the particular flavour of bug. In these
instances, the messages were being reported with the error code
SVN_ERR_WC_CORRUPT, but the bugs being detected were of the "incomplete
implementation" or "in-memory corruption" varieties. This change also
reduces the burden of translation (localization).

* subversion/libsvn_wc/tree_conflicts.c
  (svn_wc__write_tree_conflicts_to_entry): Replace error messages with
    SVN_ERR_ASSERT/SVN_ERR_MALFUNCTION.

* subversion/svn/tree-conflicts.c
  (svn_cl__append_human_readable_tree_conflict_description,
   svn_cl__append_tree_conflict_info_xml): Replace error messages with
     SVN_ERR_ASSERT/SVN_ERR_MALFUNCTION.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@873906 13f79535-47bb-0310-9956-ffa450edef68