A fix for a tree conflict problem reported by Mark Phippard.
If update/switch caused a tree conflict by trying to delete a modified
item, the working copy became unusable.  The user couldn't svn
add/cp/rm/up/ci at all.  And svn resolved didn't help, because it
restored the original state that triggered the tree conflict.

This commit changes the update editor so that an incoming delete
action will delete a tree conflict victim (and items inside it),
instead of skipping the deletion.

Such a victim, or the modified items inside it, can be reinstated via
svn add/cp.  Or the user can accept the incoming change ("theirs
full") via "rm" or "rm -rf".

The email thread for the bug report starts at:

  http://svn.haxx.se/dev/archive-2008-10/1090.shtml


* subversion/libsvn_wc/update_editor.c
  (do_entry_deletion): Don't bail out after recording a tree conflict.
   Include tree-conflict status in deletion notification.

* subversion/tests/cmdline/update_tests.py
  (update_delete_modified_files,
   update_deleted_missing_dir,
   another_hudson_problem,
   tree_conflicts_on_update_1_2,
   tree_conflicts_on_update_2_1,
   tree_conflicts_on_update_2_2,
   tree_conflicts_on_update_3): Expect update to delete tree conflict
   victims and items inside them.

* subversion/tests/cmdline/switch_tests.py
  (failed_anchor_is_target,
   tree_conflicts_on_switch_1_2,
   tree_conflicts_on_switch_2_1,
   tree_conflicts_on_switch_2_2,
   tree_conflicts_on_switch_3): Expect update to delete tree conflict
   victims and items inside them.

* subversion/tests/cmdline/commit_tests.py
  (tree_conflicts_resolved): Expect update to delete tree conflict
   victims and items inside them.

* subversion/tests/cmdline/revert_tests.py
  (revert_tree_conflicts_in_updated_files): Expect update to delete
   tree conflict victims and items inside them.

* subversion/tests/cmdline/stat_tests.py
  (status_with_tree_conflicts): Expect update to delete tree conflict
   victims and items inside them.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@874057 13f79535-47bb-0310-9956-ffa450edef68