When creating a dump file, include the sha1 checksum as well as the md5.
We don't actually parse or use the sha1 at this point, thought we could.  The
md5 is used at the repos layer to check corruption, and later passed to the fs
layer for the same reason.  If we ever switch the entire system over to sha1,
this patch will mean that 1.6-era dumpfiles will have sha1 data usable by
future versions of Subversion.

* subversion/include/svn_repos.h
  (SVN_REPOS_DUMPFILE_TEXT_COPY_SOURCE_MD5,
   SVN_REPOS_DUMPFILE_TEXT_COPY_SOURCE_SHA1,
   SVN_REPOS_DUMPFILE_TEXT_CONTENT_MD5,
   SVN_REPOS_DUMPFILE_TEXT_CONTENT_SHA1,
   SVN_REPOS_DUMPFILE_TEXT_DELTA_BASE_MD5
   SVN_REPOS_DUMPFILE_TEXT_DELTA_BASE_SHA1): New.
  (SVN_REPOS_DUMPFILE_TEXT_COPY_SOURCE_CHECKSUM,
   SVN_REPOS_DUMPFILE_TEXT_CONTENT_CHECKSUM,
   SVN_REPOS_DUMPFILE_TEXT_DELTA_BASE_CHECKSUM): Use the md5 define.

* subversion/libsvn_repos/dump.c
  (dump_node): Include the sha1 in the dumpfile.

[ in subversion/tests/cmdline/svnsync_tests_data/ ]
Fix up assorted dump files to include the now-expected sha1 field.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@874426 13f79535-47bb-0310-9956-ffa450edef68