Fix issue 2220: Deleting a non-existent property should produce an error.
The solution is a bit tricky, since our notification functions (where we
actually detect the error) don't allow us to return an error.  So, in the case
where a 'svn propdel' happens recursively, all the operations go through, and
we then send the error.  It isn't perfect, but it's the best we've got under
the circumstances.

* subversion/tests/cmdline/prop_tests.py
  (perms_on_symlink, invalid_propnames): Update to expect errors.
  (delete_nonexistent_property): New.
  (test_list): Add new test.

* subversion/svn/propdel-cmd.c
  (notify_wrapper_baton, notify_wrapper): New.
  (svn_cl__propdel): Use the new wrapper to detect a 'deleted nonexistent'
    condition, and ultimately report it if one exists.

* subversion/svn/notify.c
  (notify): Don't notify on a deleted nonexistent event.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@875310 13f79535-47bb-0310-9956-ffa450edef68